Crate namada_account

Support for signature based authorization of actions on a user account using public key(s) and signature threshold (minimum number of signatures needed to authorize an action) stored on-chain.

Modules

common

Cryptographic keys

Structs

Account

Account data

AccountPublicKeysMap

Holds the public key map data as a bimap for efficient querying

Hash

A hash, typically a sha-2 hash of a tx

InitAccount

A tx data type to initialize a new established account

Key

A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

UpdateAccount

A tx data type to update an account’s validity predicate

Enums

Address

An account’s address

Functions

clear_public_keys

Clear the public keys account subtorage space

exists

Check if a user account exists in storage

get_protocol_public_keys_key_at_addr

get_public_keys_key_at_addr

get_threshold_key_at_addr

init_account_storage

Init the subspace of a new account

is_pks_key

Check if the given storage key is a public key. If it is, returns the owner.

is_protocol_pk_key

Check if the given storage key is a public key. If it is, returns the owner.

is_protocol_public_keys_key_at_addr

is_public_keys_key_at_addr

is_threshold_key

Check if the given storage key is a threshol key.

is_threshold_key_at_addr

pks_handle

LazyMap handler for the user’s public key subspace

pks_key_prefix

Obtain a storage key for user’s public key.

protocol_pk_key

Obtain a storage key for user’s protocol public key.

public_keys

Get the public keys associated with an account

public_keys_index_map

Get the public key index map associated with an account

reveal_pk

Reveal a PK of an implicit account - the PK is written into the storage of the address derived from the PK.

set_public_key_at

Set public key at specific index

threshold

Get the threshold associated with an account

threshold_key

Obtain the storage key for a user threshold